backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
alembic==1.18.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.1
asgiref==3.11.0
asteroid-filterbanks==0.4.0
asttokens==3.0.1
attrs==25.4.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
colorlog==6.10.1
decorator==5.2.1
Django==5.0
django-cors-headers==4.3.0
djangorestframework==3.14.0
einops==0.8.1
elevenlabs==1.2.0
executing==2.2.1
ffmpeg-python==0.2.0
filelock==3.20.3
frozenlist==1.8.0
fsspec==2026.1.0
future==1.0.0
greenlet==3.3.0
gunicorn==23.0.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.3.1
HyperPyYAML==1.2.3
idna==3.11
ipython==9.9.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
julius==0.2.7
lightning==2.6.0
lightning-utilities==0.15.2
llvmlite==0.46.0
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mdurl==0.1.2
more-itertools==10.8.0
mpmath==1.3.0
multidict==6.7.0
networkx==3.6.1
numba==0.63.1
numpy==2.3.5
omegaconf==2.3.0
openai-whisper==20250625
optuna==4.6.0
packaging==25.0
pandas==2.3.3
parso==0.8.5
primePy==1.3
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.4
pure_eval==0.2.3
pyannote-core==6.0.1
pyannote-database==6.1.1
pyannote-metrics==4.0.0
pyannote-pipeline==4.0.0
pyannote.audio==3.1.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.21
pytorch-lightning==2.6.0
pytorch-metric-learning==2.9.0
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
rich==14.2.0
ruamel.yaml==0.18.17
ruamel.yaml.clib==0.2.15
scikit-learn==1.8.0
scipy==1.17.0
semver==3.0.4
sentencepiece==0.2.1
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sortedcontainers==2.4.0
soundfile==0.13.1
speechbrain==1.0.3
SQLAlchemy==2.0.45
sqlparse==0.5.5
stack-data==0.6.3
sympy==1.14.0
tensorboardX==2.6.4
threadpoolctl==3.6.0
tiktoken==0.12.0
torch==2.9.1
torch-audiomentations==0.12.0
torch_pitch_shift==1.2.5
torchaudio==2.9.1
torchmetrics==1.8.2
tqdm==4.67.1
traitlets==5.14.3
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
wcwidth==0.2.14
websockets==16.0
yarl==1.22.0
